+++ CHANGELOG.md
@@ -1,6 +1,36 @@
-# Change Log
+# Changelog
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+
+## [0.1.0](https://github.com/stewwan/capacitor-intercom/compare/v0.0.5...v0.1.0) (2019-08-23)
+
+
+### Features
+
+* add setUserHash resolves [#2](https://github.com/stewwan/capacitor-intercom/issues/2) ([8614592](https://github.com/stewwan/capacitor-intercom/commit/8614592))
+
+
+
+<a name="0.0.5"></a>
+## [0.0.5](https://github.com/stewwan/capacitor-intercom/compare/v0.0.4...v0.0.5) (2019-05-03)
+
+
+### Bug Fixes
+
+* **android:** add app compat ([9e2a97c](https://github.com/stewwan/capacitor-intercom/commit/9e2a97c))
+
+
+
+<a name="0.0.4"></a>
+## [0.0.4](https://github.com/stewwan/capacitor-intercom/compare/v0.0.3...v0.0.4) (2019-05-03)
+
+
+### Bug Fixes
+
+* **android:** cap implementation ([081eb00](https://github.com/stewwan/capacitor-intercom/commit/081eb00))
+* **README:** typo ([1898e9e](https://github.com/stewwan/capacitor-intercom/commit/1898e9e))
+
+
 
 <a name="0.0.3"></a>
 ## [0.0.3](https://github.com/stewwan/capacitor-intercom/compare/v0.0.2...v0.0.3) (2019-03-31)